Correctly handle "N/A" memory usage value
Some platforms don't report SRAM usage. It's also possible for a platform to not report flash usage, though I haven't seen a case of this. When the compile-examples action is not able to determine memory usage, it records it as "N/A".

The code previously assumed that the memory usage data would always be a number, and thus did not produce valid JSON, which requires strings to be quoted.