Handle Google API rate limiting
The Google API has a rate limit of 100 API requests per 100 seconds. The increase in requests caused by reporting for all sketches, and expanding the sheet when new columns or rows are added (mostly only significant on sheet initialization) makes it much easier to hit this limit.

This requires catching the exception caused by exceeding the rate limit, then waiting 100 s for the API request allowance to reset.