Format the documentation to make default input values easy to find
Previously, the information on whether an input had a default value -- and if so, what -- was incorporated into the input description.
This meant you had to read the entire description to find this information.

When using other actions, I have regularly had difficulty determining which inputs are required and default input values.
I would like for the users of this action to not have that experience.