Split Google API credentials setup instructions into discrete parts
Even though they are simple and easy to follow the visual length of the 39 steps of the instructions might seem daunting to someone getting started with the arduino/report-size-trends action. These are actually three discrete but interrelated operations:

1. Set up Google API credentials
2. Define GitHub repository secret
3. Set up Google Sheets spreadsheet

Splitting them into parts makes them more readable and easier to understand the purpose of each step.